Updated scripts to add json contents to dashboard db and read from there when available. Moved header check functions to make them importable for dashboard
